// src/types/infoChapter.ts
export interface InfoChapter {
  chapter: Chapter;
}

interface Chapter {
  id: number;
  revelation_place: string;
  revelation_order: number;
  bismillah_pre: boolean;
  name_simple: string;
  name_complex: string;
  name_arabic: string;
  verses_count: number;
  pages: number[];
  translated_name: TranslatedName;
}

interface TranslatedName {
  language_name: string;
  name: string;
}